+++ tests/plugins/service/test_files/loop.php
@@ -1,0 +1,6 @@
+<?php
+for ($x = 0; $x <= 1000; $x++) {
+  sleep(1);
+  error_log("The number is: $x", 0);
+}
+?>